[TASK] use and extend Files Utility to get blob list
This changes the readDirectoryRecursively so that it allows me to skip symlinks
and pass in a custom filter function, so that I can filter with regex or glob or
whatever makes sense within BlobQuery, without requiring that other people do the
same. Hopefully this updated method makes it into Flow at some point.