[TASK] Filter by mediaType using filename suffixes
This uses the extensions registered in MediaTypes to find files of a given media type.
It also adds a local copy of MediaTypes to add some important media types such as
php, sass, less, and typoscript (amongst others). Eventually, this updated version
of MediaTypes should be incorporated into TYPO3.Flow.