[TASK] Add dependency on Symfony/Finder
The code to search for files (in BlobQuery) smelled awful. After some googling and reading a
lot of code, I decided to use Symfony/Finder because it has such a clean interface and is
well organized and well documented.

One thing it does not support is using globs for the paths. However, Finder has a Glob class
which will allow me to convert the globs to regex before using them in path().

Plus, I get the benefit of using the native 'find' command on Linux and Mac. Windows is stuck
with the slower, but functional, PHP-based adapter.

Version 2.3 is required because it has ignoreUnreadableDirs() and error handling is more
consistent. Also, version 2.2 introduced path() and support for globs in in().